+++ Formatting/Posh/Posh.RSS.FeedCollection.format.ps1
@@ -1,0 +1,17 @@
+Write-FormatView -TypeName Posh.RSS.FeedCollection -Action {
+    Write-FormatViewExpression -ScriptBlock {
+        Show-Markdown -InputObject $_.README
+    }
+
+    Write-FormatViewExpression -Newline
+
+    Write-FormatViewExpression -Style 'Foreground.Cyan' -ScriptBlock {
+        @(
+        "Feeds: "
+        ""
+        foreach ($feed in $_.All) {
+            "* $($feed.Name)"
+        }
+        ) -join [Environment]::NewLine
+    }    
+}