+++ Types/PSModuleInfo/get_News.ps1
@@ -1,0 +1,25 @@
+<#
+.SYNOPSIS
+    Gets a Module's News
+.DESCRIPTION
+    Lists RSS NewsFeeds from this module
+.EXAMPLE
+    $posh.News
+#>
+
+if (-not $this.'.News') {
+    $this | 
+        Add-Member NoteProperty '.News' @(
+            $this.LinkList(@(
+                $this.PrivateData.RSS
+                $this.PrivateData.News
+                $this.PrivateData.Feed
+                $this.PrivateData.PSData.RSS
+                $this.PrivateData.PSData.News
+                $this.PrivateData.PSData.Feed
+            ), "Posh.RSS.Feed", "Posh.RSS.FeedCollection") | 
+                Add-Member NoteProperty Module $this -Force -PassThru
+        ) -Force       
+}
+
+$this.'.News'